requirements.txt
setuptools>=51.0.0
wheel


streamlit==1.31.1
plotly==5.3.1
pandas
#pandas==1.3.3
numpy
#numpy==1.21.5
pillow==8.3.2
streamlit_folium==0.3.0
folium==0.14.0
requests
matplotlib
#matplotlib==3.4.3
windrose

openmeteo_requests
retry_requests
requests_cache

pytz
